+++ src/Juegos/RuletaNumeros/Pares.java
@@ -10,7 +10,7 @@ package Juegos.RuletaNumeros;
  */
 public class Pares extends UtileriasRuleta{
     @Override
-    public int calcularGanancia() {
+    public double calcularGanancia() {
         return getApuesta() * 2; 
     }
